:recycle: refactor: make db edge compatible
