+++ src/middleware.ts
@@ -2,7 +2,7 @@ import { clerkMiddleware, createRouteMatcher } from '@clerk/nextjs/server';
 import { NextResponse } from 'next/server';
 
 import { authEnv } from '@/config/auth';
-import { auth } from '@/libs/next-auth';
+import { edgeMiddleware } from '@/libs/next-auth';
 
 import { OAUTH_AUTHORIZED } from './const/auth';
 
@@ -20,7 +20,7 @@ export const config = {
 
 const defaultMiddleware = () => NextResponse.next();
 
-const nextAuthMiddleware = auth((req) => {
+const nextAuthMiddleware = edgeMiddleware((req) => {
   // skip the '/' route
   if (req.nextUrl.pathname === '/') return NextResponse.next();
