+++ src/server/context.ts
@@ -1,13 +1,15 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import { getAuth } from '@clerk/nextjs/server';
+import { User } from 'next-auth';
 import { NextRequest } from 'next/server';
 
-import { JWTPayload, LOBE_CHAT_AUTH_HEADER, enableClerk } from '@/const/auth';
+import { JWTPayload, LOBE_CHAT_AUTH_HEADER, enableClerk, enableNextAuth } from '@/const/auth';
+import { auth as getNextAuth } from '@/libs/next-auth';
 
 type ClerkAuth = ReturnType<typeof getAuth>;
 
 export interface AuthContext {
-  auth?: ClerkAuth;
+  auth?: ClerkAuth | User;
   authorizationHeader?: string | null;
   jwtPayload?: JWTPayload | null;
   userId?: string | null;
@@ -18,7 +20,7 @@ export interface AuthContext {
  * This is useful for testing when we don't want to mock Next.js' request/response
  */
 export const createContextInner = async (params?: {
-  auth?: ClerkAuth;
+  auth?: ClerkAuth | User;
   authorizationHeader?: string | null;
   userId?: string | null;
 }): Promise<AuthContext> => ({
@@ -45,6 +47,14 @@ export const createContext = async (request: NextRequest): Promise<Context> => {
     auth = getAuth(request);
 
     userId = auth.userId;
+  }
+
+  if (enableNextAuth) {
+    const session = await getNextAuth();
+    if (session && session?.user?.id) {
+      auth = session?.user;
+      userId = session.user.id;
+    }
   }
 
   return createContextInner({ auth, authorizationHeader: authorization, userId });
